Fix errors/warnings
